feat(minify): mem-sim draft, cell-reset sub-minifiers
This adds the base code for the memory simulator.
Since it's not ready for use,
I've added 2 sub-minifiers that approximate its behavior; they are:
- mod 256 opt
- mutual-cancel opt

I'm using `regexp` because I was too lazy to implement a parser, lol.
This is sub-optimal because of multi-pass.

Since mem-sim does static-analysis of
[basic blocks](https://en.wikipedia.org/wiki/Basic_block),
cell-reset-loop sub-minifiers are a "necessary evil" (actually a hack)

Also fix utils doc-comments by moving